[EmbeddingAPI] Move XWalkSettingTest to new v6 test
-Create new v6 test & modify related xml files
-Cover the XWalkActivity and XWalkInitializer two ways

Impacted tests(approved): new 0, update 12, delete 0
Unit test platform: [Android]
Unit test result summary: pass 12, fail 0, block 0
