+++ embeddingapi/embedding-api-android-tests/suite.json
@@ -6,7 +6,7 @@
     "inst.*.py"
   ],
   "pkg-list": {
-    "embeddingapi, embeddingapiv5": {
+    "embeddingapi, embeddingapiv6": {
       "blacklist": [
         "*"
       ],
@@ -17,7 +17,8 @@
         "tests_v2.xml": "tests_v2.xml",
         "tests_v3.xml": "tests_v3.xml",
         "tests_v4.xml": "tests_v4.xml",
-        "tests_v5.xml": "tests_v5.xml"
+        "tests_v5.xml": "tests_v5.xml",
+        "tests_v6.xml": "tests_v6.xml"
       },
       "subapp-list": {
         "embeddingapiv1": {
@@ -28,7 +29,8 @@
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -49,7 +51,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -70,7 +73,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -91,7 +95,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -112,7 +117,30 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
-            "src/org/xwalk/embedding/test/v4"
+            "src/org/xwalk/embedding/test/v4",
+            "src/org/xwalk/embedding/test/v6"
+          ],
+          "copylist": {
+            "../build.gradle": "build.gradle",
+            "../libraries.gradle": "libraries.gradle",
+            "../pom.xml": "pom.xml",
+            "libs/chromium/base_java_test_support.jar": "libs/base_java_test_support.jar",
+            "libs/chromium/content_java_test_support.jar": "libs/content_java_test_support.jar",
+            "libs/chromium/net_java_test_support.jar": "libs/net_java_test_support.jar",
+            "libs/testkit/testkit-junit.jar": "libs/testkit-junit.jar"
+          },
+          "embeddingapi-library-name": "crosswalk-webview"
+        },
+        "embeddingapiv6": {
+          "app-dir": "embeddingapi",
+          "app-name": "embeddingapi_v6",
+          "app-type": "EMBEDDINGAPI",
+          "blacklist": [
+            "src/org/xwalk/embedding/test/v1",
+            "src/org/xwalk/embedding/test/v2",
+            "src/org/xwalk/embedding/test/v3",
+            "src/org/xwalk/embedding/test/v4",
+            "src/org/xwalk/embedding/test/v5"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -145,7 +173,8 @@
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -179,7 +208,8 @@
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -200,7 +230,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -235,7 +266,8 @@
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -256,7 +288,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -277,7 +310,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -313,7 +347,8 @@
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -334,7 +369,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v3",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -355,7 +391,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v4",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
@@ -376,7 +413,8 @@
             "src/org/xwalk/embedding/test/v1",
             "src/org/xwalk/embedding/test/v2",
             "src/org/xwalk/embedding/test/v3",
-            "src/org/xwalk/embedding/test/v5"
+            "src/org/xwalk/embedding/test/v5",
+            "src/org/xwalk/embedding/test/v6"
           ],
           "copylist": {
             "../build.gradle": "build.gradle",
