+++ embeddingapi/embedding-api-android-tests/tests.full.xml
@@ -113,9 +113,9 @@
           <test_script_entry>org.xwalk.embedding.test.v5.SetNetworkAvailableTest</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewSettingTest" platform="android" priority="P1" purpose="Check if the methods of XWalkSetting interface can be executed correctly." status="approved" type="functional_positive" subcase="6">
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkSettingTest" platform="android" priority="P1" purpose="Check if the methods of XWalkSetting interface can be executed correctly." status="approved" type="functional_positive" subcase="6">
         <description>
-          <test_script_entry>org.xwalk.embedding.test.v5.XWalkViewSettingTest</test_script_entry>
+          <test_script_entry>org.xwalk.embedding.test.v6.XWalkSettingTest</test_script_entry>
         </description>
       </testcase>
     </set>
