+++ embeddingapi/embedding-api-android-tests/tests_v5.xml
@@ -33,11 +33,6 @@
           <test_script_entry>org.xwalk.embedding.test.v5.SetNetworkAvailableTest</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewSettingTest" purpose="Check if the methods of XWalkSetting interface can be executed correctly."  subcase="6">
-        <description>
-          <test_script_entry>org.xwalk.embedding.test.v5.XWalkViewSettingTest</test_script_entry>
-        </description>
-      </testcase>
     </set>
   </suite>
 </test_definition>
