+++ embeddingapi/embedding-api-android-tests/tests_v6.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?xml-stylesheet type="text/xsl" href="./testcase.xsl"?>
+<test_definition>
+  <suite name="webapi-embeddingapi-xwalk-tests" category="Android embedding APIs">
+    <set name="EmbeddingApiTest" type="androidunit" location="device">
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkSettingTest" purpose="Check if the methods of XWalkSetting interface can be executed correctly."  subcase="6">
+        <description>
+          <test_script_entry>org.xwalk.embedding.test.v6.XWalkSettingTest</test_script_entry>
+        </description>
+      </testcase>
+    </set>
+  </suite>
+</test_definition>
