+++ embeddingapi/embedding-asyncapi-android-tests/tests.full.xml
@@ -113,9 +113,9 @@
           <test_script_entry>org.xwalk.embedding.test.v5.SetNetworkAvailableTestAsync</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkSettingTestAsync" platform="android" priority="P1" purpose="Check if the methods of XWalkSetting interface can be executed correctly." subcase="6">
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkSettingTestAsync" platform="android" priority="P1" purpose="Check if the methods of XWalkSetting interface can be executed correctly." subcase="6">
         <description>
-          <test_script_entry>org.xwalk.embedding.test.v5.XWalkSettingTestAsync</test_script_entry>
+          <test_script_entry>org.xwalk.embedding.test.v6.XWalkSettingTestAsync</test_script_entry>
         </description>
       </testcase>
     </set>
