+++ embeddingapi/embedding-asyncapi-android-tests/tests_v5.xml
@@ -33,11 +33,6 @@
           <test_script_entry>org.xwalk.embedding.test.v5.SetNetworkAvailableTestAsync</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkSettingTestAsync" purpose="Check if the methods of XWalkSetting interface can be executed correctly." subcase="6">
-        <description>
-          <test_script_entry>org.xwalk.embedding.test.v5.XWalkSettingTestAsync</test_script_entry>
-        </description>
-      </testcase>
     </set>
   </suite>
 </test_definition>
